move apiclient, errors and urls to ui-api
