+++ scm-ui/ui-api/src/errors.test.ts
